fix(StateFileGenerator): Move state and actions to vuex store
