+++ src/stores/mutation-types.js
@@ -5,6 +5,7 @@ const GLOBAL_AXIS_TYPE = 'SET_GLOBAL_AXIS_TYPE';
 
 const SHOW_APP = 'SHOW_APP';
 const SHOW_LANDING = 'SHOW_LANDING';
+const SAVING_STATE = 'SAVING_STATE';
 
 const VIEWS_SWAP_ORDER = 'VIEWS_SWAP_ORDER';
 const VIEWS_REORDER_QUAD = 'VIEWS_REORDER_QUAD';
@@ -19,6 +20,7 @@ export default {
   GLOBAL_AXIS_TYPE,
   SHOW_APP,
   SHOW_LANDING,
+  SAVING_STATE,
   VIEWS_SWAP_ORDER,
   VIEWS_REORDER_QUAD,
   SET_VIEWS,
